% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_boot.R
\name{cor_boot}
\alias{cor_boot}
\title{Bootstrapped Confidence Intervals for Correlations}
\usage{
cor_boot(x, y = NULL, use = "pairwise", method = "pearson",
  n_rep = 10000, conf = 0.95, seed = 42, n_cores = 2)
}
\arguments{
\item{x}{a numeric matrix or data frame.}

\item{y}{NULL (default) or a matrix or data frame with compatible dimensions
to x. The default is equivalent to y = x (but more efficient).}

\item{use}{an optional character string giving a method for handling missing
values. This must be (an abbreviation of) one of the strings "everything",
"all.obs", "complete.obs", or "pairwise.complete.obs"
(default). See \code{\link[stats]{cor}} for explanation of these options.}

\item{method}{a character string indicating which correlation coefficient is
to be computed. One of "pearson" (default), "kendall", or "spearman": can be
abbreviated.}

\item{n_rep}{The number of bootstrap replicates.}

\item{conf}{A scalar or vector containing the confidence level(s) of the
required interval(s). 95\% intervals used by default.}

\item{seed}{Integer used to seed the random number generator.}

\item{n_cores}{Integer indicating number of processes to use in parallel
processing. Set to 2 by default. To maximize speed, set to the number of
available CPUs.}
}
\value{
Object of class "cor_boot", a \code{\link{cor_list}} object with the
 following additional vectors:
\describe{
 \item{lower}{lower endpoint of the confidence interval}
 \item{upper}{upper endpoint of the confidence interval}
}
and the attribute \code{"CI"} indicating the size of the confidence interval being
returned.
}
\description{
Calculates the bias-corrected and accelerated (BCa) bootstrap confidence
interval for correlation coefficients.
}
\examples{
# Correlations with 95\% CI for the numeric variables from the iris data set
cor_boot(iris[,-5], n_rep = 1000)

# with 99\% CI
cor_boot(iris[,-5], n_rep = 1000, conf = 0.99)

}
\seealso{
\code{\link[boot]{boot.ci}}, \code{\link[stats]{cor}},
 \code{\link{cor_list}},
}
