% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_list.R
\name{cor_list}
\alias{cor_list}
\title{Construct a Correlation List}
\usage{
cor_list(x, y = NULL, use = "pairwise", method = "pearson")
}
\arguments{
\item{x}{a numeric matrix or data frame.}

\item{y}{NULL (default) or a matrix or data frame with compatible dimensions
to x. The default is equivalent to y = x (but more efficient).}

\item{use}{an optional character string giving a method for handling missing
values. This must be (an abbreviation of) one of the strings "everything",
"all.obs", "complete.obs", "na.or.complete", or "pairwise.complete.obs"
(default). See \code{\link[stats]{cor}} for explanation of these options.}

\item{method}{a character string indicating which correlation coefficient is
to be computed. One of "pearson" (default), "kendall", or "spearman": can be
abbreviated.}
}
\value{
A \code{cor_list} object with the following three vectors:
 \describe{
   \item{\code{x}}{variables that form the rows of the correlation matrix}
   \item{\code{y}}{variables that form the columns of the correlation matrix}
   \item{\code{coef}}{the correlation coefficient between corresponding
     elements of \code{x} and \code{y}}
   }
and the attribute "coef" indicating the statistic that is being returned
(Pearson's \emph{r}, Kendall's \emph{tau}, or Spearman's \emph{rho}).
}
\description{
Construct a \code{"cor_list"} object from a correlation matrix.
}
\details{
The \code{\link[stats]{cor}} function from the R stats package returns
bivariate correlations as a matrix of values, which can be difficult to parse
when there are many correlations. \code{cor_list} converts the correlation
matrix into a list of all row-column pairs, removing the diagonal entries,
and stores the type of correlation coefficient (Pearson's \emph{r}, Kendall's
\emph{tau}, or Spearman's \emph{rho}) as an attribute in a \code{cor_list}
object with specialized \code{print} and \code{summary} methods. The
\code{print} method automatically outputs unique bivariate relationships,
sorted from strongest to weakest. The \code{\link{summary.cor_list}} method
enables the use of \code{\link[dplyr]{select_helpers}} expressions to return
specific bivariate relationships that you want to look at.

\code{cor_list} differs from R's native \code{\link[stats]{cor}} in the
following ways:
\enumerate{
  \item \code{x} (and, if given, \code{y}) must be a matrix or data frame
    with named columns; unnamed vectors are not allowed. (You can still
    pass data for a single variable as \code{x} or \code{y}, but it must
    be in the form of a one-column matrix or data frame rather than a
    numeric vector.)
  \item The default argument for \code{use} is pairwise-complete observations
    instead of "everything".
  \item The \code{cor_list} object stores the type of correlation
    coefficient as an attribute.
  \item The \code{cor_list} object has special summary methods for
    selectively viewing relationships of interest. See
    \code{\link{summary.cor_list}}.
    }
}
\examples{
# Create a correlation list for the numeric variables from the iris data set
iris_cors <- cor_list(iris[,-5])

# Print returns all unique bivariate relationships sorted by strength
iris_cors

# Look at all correlations with Sepal.Length
summary(iris_cors, x = Sepal.Length)

# Look at all correlations between sepal measurements and petal measurements
summary(iris_cors, x = starts_with("Sepal"), y = starts_with("Petal"))

# Look at all correlations with Sepal.Length, excluding Sepal.Width
summary(iris_cors, x = Sepal.Length, y = -Sepal.Width)

# Look at all correlations in their original order
summary(iris_cors, sort = FALSE)

}
\seealso{
\code{\link{cor_boot}}, \code{\link{summary.cor_list}}
}
