% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.cor_list}
\alias{summary.cor_list}
\title{Summarize Correlation List}
\usage{
\method{summary}{cor_list}(object, x = everything(), y = everything(),
  sort = TRUE, ...)
}
\arguments{
\item{object}{An object of class \code{\link{cor_list}}.}

\item{x}{Variables to include/exclude from one side of the correlation.
You can use the same specifications as in \code{\link[dplyr]{select}}. If
missing, defaults to all variables.}

\item{y}{Variables to include/exclude from the other side of the
correlation. You can use the same specifications as in
\code{\link[dplyr]{select}}. If missing, defaults to all variables.}

\item{sort}{Logical value indicating whether the output should be sorted. If
\code{TRUE} (the default), output will be sorted by the absolute value of
the correlation coefficients (in descending order). If \code{FALSE}, output
will appear in the same order as the columns of the matrix or data frame
passed to \code{\link{cor}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary method for the \code{cor_list} class.
}
\details{
The \code{summary} method for the \code{cor_list} class allows for
interactive exploration of correlation results. Use
\code{\link[dplyr]{select_helpers}} expressions to return specific
bivariate relationships that you want to look at.
}
\examples{
# Create a correlation list for the numeric variables from the iris data set
iris_cors <- cor_list(iris[,-5])

# Look at all correlations with Sepal.Length
summary(iris_cors, x = Sepal.Length)

# Look at all correlations between sepal measurements and petal measurements
summary(iris_cors, x = starts_with("Sepal"), y = starts_with("Petal"))

# Look at all correlations with Sepal.Length, excluding Sepal.Width
summary(iris_cors, x = Sepal.Length, y = -Sepal.Width)

# Look at all correlations in their original order
summary(iris_cors, sort = FALSE)

}
\seealso{
\code{\link{cor_boot}}, \code{\link{cor_list}},
  \code{\link[dplyr]{select}}, \code{\link[dplyr]{select_helpers}},
}
